Move SelfTreat class to correct package
Changes:
- SelfTreatHealthProblem & SelfTreatHealthProblemMeta moved to person.health.task
- Would not work in previous commit

Part of #1315
