+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -37,7 +37,6 @@ import com.mars_sim.core.person.ai.task.meta.ReviewJobReassignmentMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewMissionPlanMeta;
 import com.mars_sim.core.person.ai.task.meta.SalvageBuildingMeta;
 import com.mars_sim.core.person.ai.task.meta.SalvageGoodMeta;
-import com.mars_sim.core.person.ai.task.meta.SelfTreatHealthProblemMeta;
 import com.mars_sim.core.person.ai.task.meta.SleepMeta;
 import com.mars_sim.core.person.ai.task.meta.TeachMeta;
 import com.mars_sim.core.person.ai.task.meta.WorkoutMeta;
@@ -49,6 +48,7 @@ import com.mars_sim.core.person.health.task.ExamineBodyMeta;
 import com.mars_sim.core.person.health.task.PrescribeMedicationMeta;
 import com.mars_sim.core.person.health.task.RequestMedicalTreatmentMeta;
 import com.mars_sim.core.person.health.task.RestingMedicalRecoveryMeta;
+import com.mars_sim.core.person.health.task.SelfTreatHealthProblemMeta;
 import com.mars_sim.core.person.health.task.TreatMedicalPatientMeta;
 import com.mars_sim.core.robot.ai.task.ChargeMeta;
 import com.mars_sim.core.robot.ai.task.SavePowerMeta;
